Refactor workers into separate modules
